# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SpeechMirror API
  version: 0.1.0
  description: 言镜三端客户端共同使用的接口契约。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      operationId: register
      security: []
      requestBody: {$ref: '#/components/requestBodies/Credentials'}
      responses:
        '200': {$ref: '#/components/responses/Tokens'}
...
  /auth/refresh:
    post:
      operationId: refreshToken
      security: []
      requestBody: {$ref: '#/components/requestBodies/RefreshToken'}
      responses:
        '200': {$ref: '#/components/responses/Tokens'}
...
  /projects:
    get:
      operationId: listProjects
      responses:
        '200': {description: 项目列表, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Project'}}}}}
    post:
      operationId: createProject
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectInput'}}}}
      responses:
        '200': {$ref: '#/components/responses/Project'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
  requestBodies:
    Credentials: {required: true, content: {application/json: {schema: {type: object, required: [username, password], properties: {username: {type: string, minLength: 3, maxLength: 32}, password: {type: string, minLength: 8, format: password}}}}}}
    RefreshToken: {required: true, content: {application/json: {schema: {type: object, required: [refresh_token], properties: {refresh_token: {type: string}}}}}}
...
  responses:
...
    Tokens: {description: 访问及刷新令牌, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}
    Project: {description: 项目, content: {application/json: {schema: {$ref: '#/components/schemas/Project'}}}}
...
  schemas:
...
    TokenPair: {type: object, required: [access_token, refresh_token, token_type, expires_in], properties: {access_token: {type: string}, refresh_token: {type: string}, token_type: {type: string, const: Bearer}, expires_in: {type: integer, const: 900}}}
    ProjectInput: {type: object, required: [name], properties: {name: {type: string, minLength: 1, maxLength: 80}, description: {type: [string, 'null']}, defense_duration_seconds: {type: integer, minimum: 60, maximum: 1800, default: 300}}}
...
    Project: {allOf: [{$ref: '#/components/schemas/ProjectInput'}, {type: object, required: [id, created_at, updated_at, defense_duration_seconds], properties: {id: {type: string, format: uuid}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}]}
...